NXT-100 Create chapter menu component and populate with urls with correct structure. Correct remaining refactored property from _links to links
