+++ web/src/feeds/publications/publications.ts
@@ -44,7 +44,7 @@ export const getAllProducts = async (): Promise<ProductLite[]> =>
 					// This means we're essentially storing a ProductLite in cache rather than a ProductLiteRaw.
 					// In perf tests this saved ~30% off the cache load time from the file system (once you factor in deserialization, file access times etc).
 					delete (product as Partial<Mutable<typeof product>>).ETag;
-					delete (product as Partial<Mutable<typeof product>>)._links;
+					delete (product as Partial<Mutable<typeof product>>).links;
 					return product;
 				}
 			)
